bait/requirements.txt
Pillow
piexif
python-telegram-bot
python-dotenv
requests
